Review and document the seed Plutono dashboard test coverage
This commit documents the test coverage
for the seed Plutono dashboard provisioning logic.
The diff is a minor improvement that I noticed while reviewing the code.

The seed Plutono is instantiated in `pkg/gardenlet/controller/seed/seed/components.go`.

The `NewPlutono` function has six arguments relevant for dashboard provisioning.
In this scenario,
five are hard coded and
only one boolean parameter depends on the seed configuration:

    shared.NewPlutono(
        r.SeedClientSet.Client(), r.GardenNamespace, secretsManager,

        component.ClusterTypeSeed,           // (1) clusterType

        1, authSecretName, seed.GetIngressFQDN("g-seed"), v1beta1constants.PriorityClassNameSeedSystem600,

        true,                                // (2) includeIstioDashboards
        false,                               // (3) isWorkerless
        false,                               // (4) isGardenCluster
        false,                               // (5) vpnHighAvailabilityEnabled
        v1beta1helper.SeedSettingVerticalPodAutoscalerEnabled(seed.GetInfo().Spec.Settings),
                                             // (6*) vpaEnabled

        wildcardCertName,
    )

Altogether, there are 2 possible combinations of this flag,
and they are covered in the unit test.

- Default seed cluster (VPA disabled)
- Seed cluster with VPA enabled

The implementation is in `getDashboardConfigMap()`:

  } else if p.values.ClusterType == component.ClusterTypeSeed {
    requiredDashboards = map[string]embed.FS{seedDashboardsPath: seedDashboards, commonDashboardsPath: commonDashboards}
    if !p.values.IncludeIstioDashboards {
      ignorePaths.Insert("istio")
    }
    if !p.values.VPAEnabled {
      ignorePaths.Insert("vpa")
    }
  }

The test coverage reveals that `!p.values.IncludeIstioDashboards` is not
executed, which is correct, because the NewPlutono instantiation sets
`includeIstioDashboards` always to `true`.
So we can remove that code path from the implementation.