Review and document the garden Plutono dashboard test coverage
This commit documents the current test coverage
for the garden Plutono dashboard provisioning logic.
The diff is a minor improvement that I noticed while reviewing the code.

The garden Plutono is instantiated in `pkg/operator/controller/garden/garden/components.go`.

The `NewPlutono` function has six arguments relevant for dashboard provisioning.
In this scenario,
five are hard coded and
only one boolean parameter depends on the garden configuration:

    shared.NewPlutono(
        r.RuntimeClientSet.Client(), r.GardenNamespace, secretsManager,

        component.ClusterTypeSeed,           // (1) clusterType

        1, "", "plutono-garden."+ingressDomain, v1beta1constants.PriorityClassNameGardenSystem100,

        false,                               // (2) includeIstioDashboards
        false,                               // (3) isWorkerless
        true,                                // (4) isGardenCluster
        false,                               // (5) vpnHighAvailabilityEnabled
        vpaEnabled(garden.Spec.RuntimeCluster.Settings),
                                             // (6*) vpaEnabled

        wildcardCertSecretName,
    )

Altogether, there are 2 possible combinations of this flag,
and they are covered in the unit test
under the seed cluster type context,
using a sub-context for `IsGardenCluster=true`:

- Default garden cluster (VPA disabled)
- Garden cluster with VPA enabled

The order of the Ginkgo `Context`, `BeforeEach` and `It` blocks
was inconsistent with the other tests, and that is now fixed.

  Context("Cluster type is seed", ...
    Context("Cluster is garden cluster", ...
      BeforeEach(...
      It("should successfully deploy all resources", ...  // this was at the end

      Context("with VPAEnabled=true", ...
        BeforeEach(...
        It("should successfully deploy all resources", ...