Fix: Show the VPA recommendations dashboard for the seeds
# Problem description

The Plutono component that is instantiated for the seed clusters did not provision the VPA recommendations dashboard.

# Problem analysis

The seed Plutono is instantiated in `pkg/gardenlet/controller/seed/seed/components.go` and
the value of `v1beta1helper.SeedSettingVerticalPodAutoscalerEnabled(seed.GetInfo().Spec.Settings)`
is used for the `VPAEnabled` flag of the Plutono component.

However, this is incorrect.

`pkg/apis/core/v1beta1/types_seed.go`:
> // Seed.Spec.Settings.VerticalPodAutoscaler.Enabled controls
> // whether the VPA components shall be deployed into the garden namespace in the seed cluster.
> // It is enabled by default because Gardener heavily relies on a VPA being deployed.
> // You should only disable this if your seed cluster already has another, manually/custom managed VPA deployment.

So this flag is not about whether the VPA is enabled in the seed cluster, but
rather whether the VPA components should be deployed by Gardener.

In a Gardener landscape, this flag is set to `true` in the top level seed
cluster, i.e. in the soil cluster. In managed seed clusters, that are Gardener
managed shoot clusters, the VPA is enabled in the shoot specification and is
deployed to the control plane of the shoot. So for such managed seed clusters,
this Seed.Spec.Settings.VerticalPodAutoscaler.Enabled flag is set to `false`:
the Gardenlet need not deploy a VPA into the seed cluster, because the VPA is
already deployed in the control plane of the seed cluster.

In `pkg/gardenlet/controller/seed/seed/reconciler_reconcile.go`, in the `runReconcileSeedFlow`:

        // VPA is a prerequisite.
        // If it's enabled then we deploy the CRD (and later also the related components)
        // as part of the flow.
        // However, when it's disabled then we check whether it is indeed available
        // (and fail, otherwise).
        if !vpaEnabled(seed.GetInfo().Spec.Settings) {
                if _, err := r.SeedClientSet.Client().RESTMapper().RESTMapping(schema.GroupKind{Group: "autoscaling.k8s.io", Kind: "VerticalPodAutoscaler"}); err != nil {
                        return fmt.Errorf("VPA is required for seed cluster: %s", err)
                }
        }

So the seed reconciliation flow fails if the seed is not configured to deploy
the VPA components by itself and the VPA is not available via the deployment in
the control plane.

Having a VPA, either in cluster or in the control plane, is a prerequisite for a
seed cluster.

# Solution

So we should not use the Seed.Spec.Settings.VerticalPodAutoscaler.Enabled flag
for the Plutono `VPAEnabled` configuration option, but rather in the seed
scenario, we should always set the `VPAEnabled` flag of the Plutono component to
`true`.

The unit test is adjusted to reflect this change.

The unit test in plutono_test.go is updated so that the seed scenario always
expects the VPA dashboards to be present (24 dashboards), and the VPA flag is
always set to true in the test setup. The redundant context for “w/ enabled vpa”
is removed, as it is now always enabled for seeds.

Note that the call site that instantiates the Plutono component, and the unit
test of the Plutono component are in different packages so in the current setup
we need to manually ensure that the unit test actually tests the relevant
scenario.