Fix: Show the VPA recommendations dashboard for the garden cluster
This ensures the VPA dashboards are always provisioned for the garden Plutono,
regardless of whether VPA is managed by Gardener or provided externally.

# Problem description

Based on the description in the previous commit, the `VPAEnabled` flag of the
Plutono component in the garden cluster scenario incorrectly uses the
`Garden.Spec.RuntimeCluster.Settings.VerticalPodAutoscaler.Enabled` flag.

This flag controls whether the VPA components shall be deployed into the runtime
garden cluster, and not whether the VPA is enabled in the runtime garden
cluster.

# Problem analysis

The garden Plutono is instantiated in `pkg/operator/controller/garden/garden/components.go` and
the value of `vpaEnabled(garden.Spec.RuntimeCluster.Settings)`
is used for the `VPAEnabled` flag of the Plutono component.

However, this is incorrect.

`pkg/apis/operator/v1alpha1/types_garden.go`:
> // Garden.Spec.RuntimeClusterSettings.VerticalPodAutoscaler.Enabled controls
> // whether the VPA components shall be deployed into this cluster.
> // It is true by default because the operator (and Gardener) heavily rely on a VPA being deployed.
> // You should only disable this if your runtime cluster already has another, manually/custom managed VPA deployment.
> // If this is not the case, but you still disable it, then reconciliation will fail.

So this flag is not about whether the VPA is enabled in the runtime garden cluster, but
rather whether the VPA components should be deployed by the Gardener operator.

In a Gardener landscape, this flag is set to `true` because the runtime cluster
is typically not a shoot cluster, but rather a IaaS cluster without VPA, such
that the Gardener operator is expected to deploy the VPA.
However, having understood a similar issue for the seed clusters in the previous
commit, this call site should also be fixed in the same way.

In `pkg/operator/controller/garden/garden/reconciler_reconcile.go`, in the `reconcile` function:

    // VPA is a prerequisite.
    // If it's enabled then we deploy the CRD (and later also the related components)
    // as part of the flow.
    // However, when it's disabled then we check whether it is indeed available
    // (and fail, otherwise).
    if !vpaEnabled(garden.Spec.RuntimeCluster.Settings) {
      if _, err := r.RuntimeClientSet.Client().RESTMapper().RESTMapping(schema.GroupKind{Group: "autoscaling.k8s.io", Kind: "VerticalPodAutoscaler"}); err != nil {
        return reconcile.Result{}, fmt.Errorf("VPA is required for runtime cluster but CRD is not installed: %s", err)
      }
    }

So the garden runtime cluster's reconciliation flow fails if the garden runtime
cluster is not configured to deploy the VPA components by itself and the VPA is
not available via some other means.

Having a VPA, either in cluster or in the control plane, is a prerequisite for a
garden runtime cluster.

# Solution

So we should not use the Garden.Spec.RuntimeCluster.Settings.VerticalPodAutoscaler.Enabled flag
for the Plutono `VPAEnabled` configuration option, but rather in the garden
scenario, we should always set the `VPAEnabled` flag of the Plutono component to
`true`.

The unit test is adjusted to reflect this change.

The unit test in plutono_test.go is updated so that the garden scenario always
expects the VPA dashboards to be present (26 dashboards), and the VPA flag is
always set to true in the test setup. The redundant context for “w/ enabled vpa”
is removed, as it is now always enabled for garden scenario.

Note that the call site that instantiates the Plutono component, and the unit
test of the Plutono component are in different packages so in the current setup
we need to manually ensure that the unit test actually tests the relevant
scenario.