WIP Fix: Show the VPA recommendations dashboard for the shoots
# Problem description

The Plutono component that is instantiated for the shoot clusters does not
always provision the VPA recommendations dashboard.

# Problem analysis

The shoot Plutono is instantiated in `pkg/gardenlet/operation/botanist/plutono.go` and
the value of `b.Shoot.WantsVerticalPodAutoscaler`
is used for the `VPAEnabled` flag of the Plutono component.

However, this is incorrect.

`ShootWantsVerticalPodAutoscaler` is defined in `pkg/apis/core/v1beta1/helper/shoot.go` as
`shoot.Spec.Kubernetes.VerticalPodAutoscaler.Enabled`.

`pkg/apis/core/v1beta1/types_shoot.go`:
> // Enabled specifies whether the Kubernetes VPA shall be enabled for the shoot cluster.

So this flag controls whether the VPA components should be deployed by Gardener
for the shoot cluster, into the control plane of the shoot cluster.

It is a feasible configuration to have a shoot cluster without a VPA, or another
one with a VPA.

The caveat is that the "VPA Recommendations" dashboard in Plutono shows the VPA
recommendations both for the pods in the control plane, running in the seed, and
for Gardener managed resources running as part of the workload in the shoot
cluster's kube-system namespace.

The control plane runs in the seed cluster, and the seed cluster always uses a VPA.

So it is always meaningful to show the VPA recommendations dashboard for the shoot cluster.

If the shoot VPA setting is set to false, it'll show the VPA recommendations
only for the control plane components. If it is set to true, it will show the
VPA recommendations for both the control plane and the workload pods in the
kube-system namespace.

# Solution

So we should not use the `b.Shoot.WantsVerticalPodAutoscaler` flag
for the Plutono `VPAEnabled` configuration option, but rather in the shoot
scenario, we should always set the `VPAEnabled` flag of the Plutono component to
`true`.