Refactor: adjust the shoot scenario
The 13th and 14th parameters of NewPlutono are
vpaRecommendationsAvailable, vpaInternalsAvailable.
By the way, this diff shows that the current approach with several
parameters is not so well readible in the source code at the call site.

So in the shoot scenario, Prometheus runs in the control plane of the shoot.
If the shoot is configured with
  shoot.Spec.Kubernetes.VerticalPodAutoscaler.Enabled
then the VPA components are going to be deployed in the control plane,
next to Prometheus. So Prometheus can scrape the VPA internal metrics.

Note that the VPA recommendations in the shoot scenario have 2 sources:
1: the control plane components. The control plane components running in the
seed are always under VPA.
2: the components in the shoot's kube-system namespace. They are under VPA
only if the shoot is configured to deploy a VPA.