Add health recording rules to aggregate Prometheus
Health checks are added to aggregate Prometheus to detect scrape job
failures and missing federated metering metrics.

Recording rules tests are moved to the component package to get access
to aggregate Prometheus internals. This follows the same pattern used
for garden and longterm Prometheus.

The shoot-prometheus job is excluded from health checking because issues
with shoot control planes can take it down. Such failures are reported
via SeedSystemComponentsHealthy, which blocks seed operations if it is
set to false.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>