Extend shoot care for Prometheus health checking
This adds Prometheus health checking to the shoot care controller.
Prometheus resources are listed from the API server and filtered based
on the "health-check-by" label.

This also adds an end-to-end test that deploys a failing health check
rule in shoot Prometheus and validates the corresponding conditions are
successfully set to false.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>