Add end-to-end tests for Prometheus health checking
This commit adds end-to-end tests that deploy a test PrometheusRule to
simulate a down scrape job on every Prometheus instance:

    up{job="gardener-e2e-test"}=0

and validates that the corresponding conditions are set to false.

For the seed case, the managed seed client is extended with the seed
scheme. Since managed seeds are shoots with gardenlet installed, clients
can be built using seed schemes in addition to shoot schemes. This
allows creation and deletion of PrometheusRules during end-to-end tests.
Using a managed seed for the end-to-end tests allows tearing it down
after completion. Running directly on the KinD seed would be too
intrusive.

Furthermore, two smaller changes are required for the end-to-end test
setup:

1. The garden resource as part of operator-seed-up needs to be
   reconciled for the garden operator to pick up the local seed and
   create the corresponding configurations in the garden Prometheus to
   federate from the seed's aggregate Prometheus. Without this,
   federation regressions between these two Prometheus instances would
   go undetected.
2. The local setup needs to wait for the observability components to be
   ready. If this condition does not become ready timely, the local
   setup will fail and so will the integration tests.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>